Refactor filters stuff to separate file
